Nov 14, 2024, 3:45 PM

+++ temp.sol
@@ -15,7 +15,7 @@ contract Deployer {
                 bytes32(0x0)  // Merkle root for distribution 2
             ],
             [0, 0], // Cliff periods for each distribution
-            [0, 0], // TGE percentages for each distribution
+            [uint256(0), uint256(0)], // TGE percentages for each distribution
             [0, 0]  // Total rounds for each distribution
         );
         token.Mint(address(distributor), 1000000000000000000000000000);
